requirements.txt
Django==2.0
image==1.5.17
numpy==1.13.3
olefile==0.44
opencv-python==3.4.0.12
Pillow==4.3.0
PyAutoGUI==0.9.36
PyMsgBox==1.0.6
PyScreeze==0.1.13
PyTweening==1.0.3
pytz==2017.3
